Fix dashboard clicks not working on lower sidebar items
rowIndexAt() was not checking showKeymapHints when calculating the
clickable area, but View() and visibleHeight() did check it. This caused
clicks on lower items to fail bounds checks when keymap hints were hidden
(the default).

Added helpLineCount() helper function (matching sidebar pattern) to
encapsulate the showKeymapHints check and prevent this class of bug.
Refactored rowIndexAt(), View(), and visibleHeight() to use it.

Added regression test and split toolbar tests to separate file to stay
under 500-line limit.
